refactor: Split & rename function for build store link
